requirement.txt
google-generativeai
streamlit
folium
streamlit-folium
requests
python-dotenv